Move version.readSeries to be linux-only.
