+++ version/osversion.go
@@ -4,7 +4,6 @@
 package version
 
 import (
-	"fmt"
 	"strconv"
 	"strings"
 
@@ -37,34 +36,6 @@ func MustOSFromSeries(series string) os.OSType {
 		panic("osVersion reported an error: " + err.Error())
 	}
 	return operatingSystem
-}
-
-func getValue(from map[string]string, val string) (string, error) {
-	for serie, ver := range from {
-		if ver == val {
-			return serie, nil
-		}
-	}
-	return "unknown", errors.New("Could not determine series")
-}
-
-func readSeries() (string, error) {
-	values, err := os.ReadOSRelease(osReleaseFile)
-	if err != nil {
-		return "unknown", err
-	}
-	updateSeriesVersions()
-	switch values["ID"] {
-	case strings.ToLower(os.Ubuntu.String()):
-		return getValue(ubuntuSeries, values["VERSION_ID"])
-	case strings.ToLower(os.Arch.String()):
-		return getValue(archSeries, values["VERSION_ID"])
-	case strings.ToLower(os.CentOS.String()):
-		codename := fmt.Sprintf("%s%s", values["ID"], values["VERSION_ID"])
-		return getValue(centosSeries, codename)
-	default:
-		return "unknown", nil
-	}
 }
 
 // kernelToMajor takes a dotted version and returns just the Major portion
